chore(refactor): extract /api/import/* into routes/import.ts
Seventh bucket of the index.ts route extraction. Three endpoints:
  GET  /api/import/prompt     — generate the prompt the user pastes elsewhere
  POST /api/import/preview    — convert pasted output → import plan
  POST /api/import/execute    — apply an approved plan

Two collateral changes worth flagging:

1. http-utils.readJsonBody now accepts an optional { maxBytes } cap,
   defaulting to MAX_MUTATION_BODY (64 KB). The import endpoints raise
   it (preview = 500 KB, execute = 100 KB) — they accept pasted AI
   output that's legitimately larger than typical mutation bodies.
   This kills the pre-existing duplicate body-cap parsing the audit
   flagged (item 11) and migrates both off the legacy req.on('data')
   shape.

2. MemoryProvider interface gained `findExact(content, spaceId?)`.
   The method existed on the concrete SqliteFtsMemoryProvider class,
   but index.ts was reaching past the interface to call it (TS was
   happy because index instantiated the concrete type directly).
   Adding it to the interface is the right place — the import flow
   relies on it for dedupe, and any future provider must support it.

Drive-by: dropped the now-unused `resolveSpaceRow` helper from index.ts
(moved into routes/import.ts as a local closure) and trimmed the
import.ts named imports down to `setImportStatePath` (the only one
still needed at module-init time).

Cumulative: index.ts is now 1115 lines — down 1103 lines (-50%) from
the pre-audit baseline of 2218.

No behavioural change.
